fix: resolve clippy warnings for Rust 1.93
- Remove unnecessary usize casts in meilisearch client
- Use #[derive(Default)] with #[default] attribute instead of manual impl
- Applies to Visibility, NoteStatus, NoteImportance, ViolationAction enums

+++ src/meilisearch/client.rs
@@ -266,7 +266,7 @@ impl MeiliClient {
         let index = self.client.index(index_names::CODE);
         let stats = index.get_stats().await?;
         Ok(IndexStats {
-            total_documents: stats.number_of_documents as usize,
+            total_documents: stats.number_of_documents,
             is_indexing: stats.is_indexing,
         })
     }
@@ -462,7 +462,7 @@ impl MeiliClient {
         let index = self.client.index(index_names::NOTES);
         let stats = index.get_stats().await?;
         Ok(IndexStats {
-            total_documents: stats.number_of_documents as usize,
+            total_documents: stats.number_of_documents,
             is_indexing: stats.is_indexing,
         })
     }
